// src/lib/schemas.ts
export const LastFmImage = z.object({
  "#text": z.string(),
  size: z.union([z.literal('small'), z.literal('medium'), z.literal('large'), z.literal('extralarge'), z.literal('mega'), z.literal("")]),
})
// ...
export const LastFmTopTracks = z.object({
  streamable: z.object({
    fulltrack: z.string(),
    "#text": z.string(),
  }),
  mbid: z.string(),
  name: z.string(),
  image: z.array(LastFmImage),
  artist: z.object({
    url: z.string().url(),
    name: z.string(),
    mbid: z.string(),
  }),
  url: z.string().url(),
  duration: z.string(),
  '@attr': z.object({
    rank: z.string(),
  }),
  playcount: z.string(),
})
// ...
export const SpotifyAlbumSchema = z.object({
  album_type: z.string(),
  href: z.string().url(),
  id: z.string(),
  name: z.string(),
  release_date: z.string(),
  total_tracks: z.number(),
  type: z.string(),
  images: z.array(z.object({
    height: z.number(),
    url: z.string().url(),
    width: z.number(),
  })),
})

export const SpotifyTrackSchema = z.object({
  album: SpotifyAlbumSchema,
  disc_number: z.number(),
  duration_ms: z.number(),
  explicit: z.boolean(),
  external_urls: z.object({
    spotify: z.string().url(),
  }),
  id: z.string(),
  name: z.string(),
  popularity: z.number(),
  track_number: z.number(),
  type: z.string(),
})